Use decimal types for article combo lookups

+++ src/ERPAccounting.Domain/Lookups/LookupModels.cs
@@ -61,9 +61,9 @@ public record ArticleLookup(
     [property: Column("NAZIV ARTIKLA")] string NazivArtikla,
     [property: Column("JM")] string? JedinicaMere,
     [property: Column("IDPoreskaStopa")] string? IdPoreskaStopa,
-    [property: Column("ProcenatPoreza")] double ProcenatPoreza,
-    [property: Column("Akciza")] double Akciza,
-    [property: Column("KoeficijentKolicine")] double KoeficijentKolicine,
+    [property: Column("ProcenatPoreza")] decimal ProcenatPoreza,
+    [property: Column("Akciza")] decimal Akciza,
+    [property: Column("KoeficijentKolicine")] decimal KoeficijentKolicine,
     [property: Column("ImaLot")] bool ImaLot,
     [property: Column("OtkupnaCena")] decimal? OtkupnaCena,
     [property: Column("PoljoprivredniProizvod")] bool PoljoprivredniProizvod
